requirements.txt
opencv-python
torch
torchaudio
torchvision
pillow
numpy
glob